YD-398 Processed review comments
* Spelling error in variable named
* Moved encrypt and decrypt methods to CryptoUtil
* Renamed *Encrypter to *Encryptor
