+++ core/src/main/java/nu/yona/server/messaging/entities/MessageSource.java
@@ -24,7 +24,7 @@ import org.springframework.data.domain.Pageable;
 import nu.yona.server.analysis.entities.IntervalActivity;
 import nu.yona.server.crypto.pubkey.PublicKeyDecryptor;
 import nu.yona.server.crypto.pubkey.PublicKeyUtil;
-import nu.yona.server.crypto.seckey.ByteFieldEncrypter;
+import nu.yona.server.crypto.seckey.ByteFieldEncryptor;
 import nu.yona.server.entities.EntityWithUuid;
 import nu.yona.server.entities.RepositoryProvider;
 import nu.yona.server.messaging.service.MessageNotFoundException;
@@ -38,7 +38,7 @@ public class MessageSource extends EntityWithUuid
 		return (MessageSourceRepository) RepositoryProvider.getRepository(MessageSource.class, UUID.class);
 	}
 
-	@Convert(converter = ByteFieldEncrypter.class)
+	@Convert(converter = ByteFieldEncryptor.class)
 	@Column(length = 1024)
 	private byte[] privateKeyBytes;
